+++ waves/tests/test_parameter_generator.py
@@ -1,14 +1,14 @@
 """Test ParameterGenerator Abstract Base Class
 """
 
-import pytest
 from unittest.mock import patch, mock_open
 import pathlib
 
+from waves.parameter_generators import _ParameterGenerator
+
+import pytest
 import numpy
 import xarray
-
-from waves.parameter_generators import _ParameterGenerator
 
 
 class TestParameterGenerator:
@@ -40,15 +40,18 @@ class TestParameterGenerator:
         :param bool is_file: test specific argument mocks output for pathlib.Path().is_file()
         :param int sets: test specific argument for the number of sets to build for the test
         """
-        WriteParameterGenerator = NoQuantilesGenerator(schema, template, overwrite, dryrun, debug)
+        WriteParameterGenerator = NoQuantilesGenerator(schema, output_file_template=template, output_file_type='yaml',
+                                                       overwrite=overwrite, dryrun=dryrun, debug=debug)
         WriteParameterGenerator.generate(sets)
         with patch('waves.parameter_generators._ParameterGenerator._write_meta'), \
-             patch('builtins.open', mock_open(read_data='schema')) as mock_file, \
+             patch('builtins.open', mock_open()) as mock_file, \
              patch('sys.stdout.write') as stdout_write, \
+             patch('xarray.Dataset.to_netcdf') as xarray_to_netcdf, \
              patch('pathlib.Path.is_file', side_effect=is_file), \
              patch('pathlib.Path.mkdir'):
             WriteParameterGenerator.write()
             mock_file.assert_not_called()
+            xarray_to_netcdf.assert_not_called()
             assert stdout_write.call_count == sets
 
     init_write_files = {# schema, template, overwrite, dryrun, debug,          is_file, sets, files
@@ -66,7 +69,7 @@ class TestParameterGenerator:
                                  init_write_files.values(),
                              ids=init_write_files.keys())
     def test_write_to_files(self, schema, template, overwrite, dryrun, debug, is_file, sets, files):
-        """Check for conditions that should result in calls to stdout
+        """Check for conditions that should result in calls to builtins.open
 
         :param str schema: placeholder string standing in for the schema read from an input file
         :param str template: user supplied string to be used as a template for output file names
@@ -75,17 +78,50 @@ class TestParameterGenerator:
         :param bool debug: print cli debugging information and exit
         :param list is_file: test specific argument mocks changing output for pathlib.Path().is_file() repeat calls
         :param int sets: test specific argument for the number of sets to build for the test
+        :param int files: integer number of files that should be written
         """
-        WriteParameterGenerator = NoQuantilesGenerator(schema, template, overwrite, dryrun, debug)
+        WriteParameterGenerator = NoQuantilesGenerator(schema, output_file_template=template, output_file_type='yaml',
+                                                       overwrite=overwrite, dryrun=dryrun, debug=debug)
         WriteParameterGenerator.generate(sets)
         with patch('waves.parameter_generators._ParameterGenerator._write_meta'), \
-             patch('builtins.open', mock_open(read_data='schema')) as mock_file, \
+             patch('builtins.open', mock_open()) as mock_file, \
              patch('sys.stdout.write') as stdout_write, \
+             patch('xarray.Dataset.to_netcdf') as xarray_to_netcdf, \
+             patch('pathlib.Path.is_file', side_effect=is_file):
+            WriteParameterGenerator.write()
+            stdout_write.assert_not_called()
+            xarray_to_netcdf.assert_not_called()
+            assert mock_file.call_count == files
+
+    @pytest.mark.unittest
+    @pytest.mark.parametrize('schema, template, overwrite, dryrun, debug, is_file, sets, files',
+                                 init_write_files.values(),
+                             ids=init_write_files.keys())
+    def test_write_to_netcdf(self, schema, template, overwrite, dryrun, debug, is_file, sets, files):
+        """Check for conditions that should result in calls to xarray.Dataset.to_netcdf
+
+        :param str schema: placeholder string standing in for the schema read from an input file
+        :param str template: user supplied string to be used as a template for output file names
+        :param bool overwrite: overwrite existing files
+        :param bool dryrun: skip file write, but show file name and associated contents that would ahve been written
+        :param bool debug: print cli debugging information and exit
+        :param list is_file: test specific argument mocks changing output for pathlib.Path().is_file() repeat calls
+        :param int sets: test specific argument for the number of sets to build for the test
+        :param int files: integer number of files that should be written
+        """
+        WriteParameterGenerator = NoQuantilesGenerator(schema, output_file_template=template, output_file_type='h5',
+                                                       overwrite=overwrite, dryrun=dryrun, debug=debug)
+        WriteParameterGenerator.generate(sets)
+        with patch('waves.parameter_generators._ParameterGenerator._write_meta'), \
+             patch('builtins.open', mock_open()) as mock_file, \
+             patch('sys.stdout.write') as stdout_write, \
+             patch('xarray.Dataset.to_netcdf') as xarray_to_netcdf, \
              patch('pathlib.Path.is_file', side_effect=is_file), \
              patch('pathlib.Path.mkdir'):
             WriteParameterGenerator.write()
+            mock_file.assert_not_called()
             stdout_write.assert_not_called()
-            assert mock_file.call_count == files
+            assert xarray_to_netcdf.call_count == files
 
     def test_create_parameter_set_names(self):
         """Test the parmater set name generation"""